Initial attempt at using toolchains with cmake.
The main rationale for this is to fully embed compiler information in
the makefiles so that environment modules do not need to be loaded at
each return to a project.

There is a bit of ambiguity as to whether the toolchains should
specify the baselibs build or if baselibs should provide the baselibs
build.  Currently doing the former.

Quite likely one does not need both an initial cache and a toolchain
file.  If just one will suffice that would be preferable.

To use:

% cmake <src-dir> -C <env>/CMakeCache.txt -DCMAKE_TOOLCHAIN_FILE=<env>/toolchain.txt