Enable more parallelism with parallel_build script
Small edits that should allow one to use most of a compute node
when running `parallel_build.csh`. Now it will use all but 2 cores for
the build.

But other edits were made so that if you are running interactively on a
*head* node you don't use, say, 26 of the 28 cores, but rather 6. This
might be too conservative, but I'd like to hear *officially* from NCCS
as to what `make -jN` they'd allow.