Update to use Baselibs 5.2.8. Use Intel 19 at NAS
This is mainly because the beta snap of GEOS in 5.2.7 had a left over
debug print from Gerhard or Walter or someone. From my testing it's
zero-diff.

As for the Intel 19 change, since Intel 18.0.5 is a broken module at
NAS, I've moved it to use that there. I've tested MAPL 2.0 with Intel 19
and baselibs 5.2.7, so I'm assuming this will work, but since NAS is
very slow, it'll be tomorrow before I can do build+run tests. But it's
not like we don't break NAS every so often