Fix order of operations with LD_LIBRARY_PATH and csh
Testing at NAS shows that for some reason, in 5.3.0, the order of operations on when
you append the Baselibs `LD_LIBRARY_PATH` to the `LD_LIBRARY_PATH` in the csh shell can
break TCL modules. So we reorder to load the modules first and then append to the `LD_LIBRARY_PATH`.